requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gspread==5.9.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
